+++ .gitignore
@@ -9,9 +9,9 @@ node_modules/
 *.egg-info/
 .ipynb_checkpoints
 *.tsbuildinfo
-jupyter_geotiler/labextension
+jupyter_tiler/labextension
 # Version file is handled by hatchling
-jupyter_geotiler/_version.py
+jupyter_tiler/_version.py
 
 # Integration tests
 ui-tests/test-results/
