+++ _jupyter-config/server-config/jupyter_tiler.json
@@ -1,7 +1,7 @@
 {
   "ServerApp": {
     "jpserver_extensions": {
-      "jupyter_geotiler": true
+      "jupyter_tiler": true
     }
   }
 }
