+++ docs/contributor-guide/index.md
@@ -7,6 +7,6 @@
 how-tos/index
 ```
 
-This documentation is aimed at contributors to `jupyter-geotiler`.
+This documentation is aimed at contributors to `jupyter-tiler`.
 
 👈 Use the navigation on the left to continue!
