+++ docs/user-guide/quickstart.md
@@ -1,11 +1,11 @@
 # Quickstart
 
 As an author of a an interactive map library for Jupyter, you might use
-`jupyter-geotiler` to provide the ability to dynamically visualize data in Xarray
+`jupyter-tiler` to provide the ability to dynamically visualize data in Xarray
 DataArrays without writing to a file like so:
 
 ```python
-from jupyter_geotiler.titiler import add_data_array
+from jupyter_tiler.titiler import add_data_array
 
 
 class MyMapLibrary:
