--- jupyter_geotiler/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Python unit tests for jupyter_geotiler."""
