+++ jupyter_tiler/constants/_jupyter.py
@@ -1,2 +1,2 @@
 # Must match the submodule name, i.e. the name you'd use to import:
-SERVER_EXTENSION_NAME = "jupyter_geotiler"
+SERVER_EXTENSION_NAME = "jupyter_tiler"
